feat: 채팅방 controller, service, domain, repo 구현완료

+++ chat-service/src/main/java/org/example/chatservice/chatMessage/infrastructure/entity/ChatMessageEntity.java
@@ -15,7 +15,7 @@ public class ChatMessageEntity {
     private  String userId;
     private String roomId;
     private String content;
-    private  int createdAt;
+    private  long createdAt;
 
     public static ChatMessageEntity from(ChatMessage chatMessage){
         return ChatMessageEntity.builder()
